+++ api/src/consts.ts
@@ -2,4 +2,5 @@ export enum SPEC_VERSION {
   V1000 = 1000,
   V1010 = 1010,
   V1100 = 1100,
+  V1200 = 1200,
 }