+++ api/src/types/payload.ts
@@ -1,5 +1,4 @@
 import { AnyJson } from '@polkadot/types/types';
 import { Bytes } from '@polkadot/types';
-import { HexString } from '@polkadot/util/types';
 
-export type PayloadType = HexString | Uint8Array | string | Bytes | AnyJson;
+export type PayloadType = Uint8Array | Bytes | AnyJson;